refactor(convert): extract opacity into convert/style/opacity.rs
